Time: 48 ms (29.90%) | Memory: 44.1 MB (55.39%) - LeetSync

+++ 2808-painting-the-walls/painting-the-walls.java
@@ -2,14 +2,14 @@ class Solution {
 
     public int paintWalls(int[] cost, int[] time) {
         int n = cost.length;
-        long[][] dp = new long[n][n];
-        for(long[] row: dp) {
+        int[][] dp = new int[n][n];
+        for(int[] row: dp) {
             Arrays.fill(row, -1);
         }
-        return (int)solve(0, 0, cost, time, dp);
+        return solve(0, 0, cost, time, dp);
     }
 
-    long solve(int i, int done, int[] cost, int[] time, long[][] dp) {
+    int solve(int i, int done, int[] cost, int[] time, int[][] dp) {
         int n = cost.length;
         if (done >= n) return 0;
 
@@ -25,9 +25,9 @@ class Solution {
         }
 
 
-        long p = cost[i] + solve(i + 1, 1 + done + time[i], cost, time, dp);
+        long p = cost[i] + (long)solve(i + 1, 1 + done + time[i], cost, time, dp);
         long up = solve(i + 1, done, cost, time, dp);
 
-        return dp[i][done] = Math.min(p, up);
+        return dp[i][done] = (int)Math.min(p, up);
     }
 }
